Los usuarios no pueden insertar, modificar o borrar los videos de otros

+++ ipartube/src/main/java/com/ipartek/formacion/ipartube/servicios/UsuarioServiceImpl.java
@@ -15,28 +15,53 @@ import lombok.extern.java.Log;
 public class UsuarioServiceImpl extends AnonimoServiceImpl implements UsuarioService {
 	@Autowired
 	private UsuarioRepository usuarioRepository;
-	
+
 	@Override
 	public Usuario obtenerUsuarioPorEmail(String email) {
 		return usuarioRepository.findByEmail(email);
 	}
 
 	@Override
-	public Video altaVideo(Video video) {
+	public Video altaVideo(String email, Video video) {
 		log.info("ALTA " + video.toString());
+
+		var usuario = usuarioRepository.findByEmail(email);
+
+		if (usuario.getId() != video.getUsuario().getId()) {
+			throw new ServiciosException("El video pertenece a un usuario que no es el que ha pedido el cambio");
+		}
+
+		return videoRepository.save(video);
+	}
+
+	@Override
+	public Video modificarVideo(String email, Video video) {
+		log.info("MODIFICAR " + video.toString());
+
+		var usuario = usuarioRepository.findByEmail(email);
+
+		if (usuario.getId() != video.getUsuario().getId()) {
+			throw new ServiciosException("El video pertenece a un usuario que no es el que ha pedido el cambio");
+		}
 		
 		return videoRepository.save(video);
 	}
 
 	@Override
-	public Video modificarVideo(Video video) {
-		log.info("MODIFICAR " + video.toString());
-	
-		return videoRepository.save(video);
-	}
+	public void bajaVideo(String email, Long id) {
+		var usuario = usuarioRepository.findByEmail(email);
+		var videoOptional = videoRepository.findById(id);
+		
+		if(videoOptional.isEmpty()) {
+			throw new ServiciosException("El video no existe");
+		}
+		
+		var video = videoOptional.get();
 
-	@Override
-	public void bajaVideo(Long id) {
+		if (usuario.getId() != video.getUsuario().getId()) {
+			throw new ServiciosException("El video pertenece a un usuario que no es el que ha pedido el cambio");
+		}
+		
 		videoRepository.deleteById(id);
 	}
